Perf(utils): Attach configs via pure Object.assign, restore co-located modules
